+++ cognite/neat/app/ui/neat-app/src/views/TransformationView.tsx
@@ -25,6 +25,7 @@ import Container from '@mui/material/Container';
 import CdfDownloader from 'components/CdfDownloader';
 import Alert from '@mui/material/Alert';
 import AlertTitle from '@mui/material/AlertTitle';
+import { Tab, Tabs } from '@mui/material';
 
 function Row(props: { row: any,properties: any }) {
   const { row,properties } = props;
@@ -98,11 +99,40 @@ function Row(props: { row: any,properties: any }) {
   );
 }
 
+/*
+"metadata": {
+        "prefix": "HqDemo",
+        "suffix": "HqDemo",
+        "namespace": "http://purl.org/cognite/neat/",
+        "version": "1.0.2",
+        "title": "HQB CIM data model",
+        "description": "This data model has been inferred with NEAT from HQB CIM dump",
+        "created": "2024-02-08T13:09:20.832000",
+        "updated": "2024-02-08T13:09:20.832000",
+        "creator": [
+            "NEAT"
+        ],
+        "contributor": [
+            "NEAT"
+        ],
+        "rights": "Unknown rights of usage",
+        "license": "Proprietary License",
+        "dataModelId": "HqDemo",
+        "source": null
+    },
+  */
+
 export default function TransformationTable() {
   const neatApiRootUrl = getNeatApiRootUrl();
-  const [data, setData] = useState({"classes":[],"properties":[],"file_name":"","hash":"","error_text":"","src":""});
+  const [data, setData] = useState({"classes":[],"properties":[],"file_name":"","hash":"","error_text":"","src":"",
+  "metadata":{"prefix":"","suffix":"","namespace":"","version":"","title":"","description":"","created":"","updated":"","creator":[],"contributor":[],"rights":"","license":"","dataModelId":"","source":""}});
   const [alertMsg, setAlertMsg] = useState("");
   const [selectedWorkflow, setSelectedWorkflow] = useState<string>(getSelectedWorkflowName());
+  const [selectedTab, setSelectedTab] = useState(1);
+
+  const handleTabChange = (event: React.SyntheticEvent, newValue: number) => {
+    setSelectedTab(newValue);
+  };
   const columns: GridColDef[] = [
     {field: 'id', headerName: 'ID', width: 70},
     {field: 'name', headerName: 'Name', width: 130},
@@ -150,29 +180,74 @@ export default function TransformationTable() {
       <AlertTitle>Warning</AlertTitle>
         {alertMsg}
     </Alert> )}
-    <TableContainer component={Paper}>
-      <Table aria-label="collapsible table">
-        <TableHead>
-          <TableRow>
-            <TableCell />
-            <TableCell> <b>Solution class</b></TableCell>
-            <TableCell align="right"><b>Description</b></TableCell>
-            <TableCell align="right"><b>CDF resource</b></TableCell>
-            <TableCell align="right"><b>Parent resource</b></TableCell>
-            <TableCell align="right"><b>Action</b></TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {data.classes.map((row:any) => (
-            <Row key={row.class} row={row} properties={data.properties} />
-          ))}
-        </TableBody>
-      </Table>
-    </TableContainer>
+        <Box>
+          <Tabs value={selectedTab} onChange={handleTabChange} aria-label="Metadata tabs">
+            <Tab label="Metadata" />
+            <Tab label="Data model" />
+          </Tabs>
+
+          {selectedTab === 0 && (
+            <Box sx={{marginTop:5}}>
+               <Typography variant="body1" gutterBottom>
+                Title: {data.metadata.title}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Description: {data.metadata.description}
+              </Typography>
+              <Typography variant="body1" gutterBottom>
+                Prefix/DM Space: {data.metadata.prefix}
+              </Typography>
+              <Typography variant="body1" gutterBottom>
+                Data Model ID: {data.metadata.dataModelId}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Namespace: {data.metadata.namespace}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Version: {data.metadata.version}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Created: {data.metadata.created}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Updated: {data.metadata.updated}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Creator: {data.metadata.creator.join(", ")}
+              </Typography>
+              <Typography variant="body2" gutterBottom>
+                Source: {data.metadata.source}
+              </Typography>
+            </Box>
+          )}
+     {selectedTab === 1 && (
+        <TableContainer component={Paper}>
+          <Table aria-label="collapsible table">
+            <TableHead>
+              <TableRow>
+                <TableCell />
+                <TableCell> <b>Solution class</b></TableCell>
+                <TableCell align="right"><b>Description</b></TableCell>
+                <TableCell align="right"><b>CDF resource</b></TableCell>
+                <TableCell align="right"><b>Parent resource</b></TableCell>
+                <TableCell align="right"><b>Action</b></TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              {data.classes.map((row:any) => (
+                <Row key={row.class} row={row} properties={data.properties} />
+              ))}
+            </TableBody>
+          </Table>
+        </TableContainer>
+     )}
+   </Box>
     <Box sx={{margin:5}}>
+      <Box sx={{width : 500}}>
+      <LocalUploader fileType="rules" action="none" stepId="none" label="Upload new data model" workflowName={selectedWorkflow} onUpload={onUpload} />
+      </Box>
       <CdfPublisher type="transformation rules" fileName={data.file_name} />
       <CdfDownloader type="neat-wf-rules" onDownloadSuccess={onDownloadSuccess} />
-      <LocalUploader fileType="rules" action="none" stepId="none" workflowName={selectedWorkflow} onUpload={onUpload} />
     </Box>